Add wait for attack set time action and move wait for attack actions to be on the same file
