+++ playerbot/PlayerbotAI.cpp
@@ -845,7 +845,7 @@ void PlayerbotAI::HandleCommand(uint32 type, const string& text, Player& fromPla
             bot->GetSession()->HandleLogoutCancelOpcode(p);
         }
     }
-    else if (filtered.size() > 5 && filtered.substr(0, 5) == "wait ")
+    else if ((filtered.size() > 5) && (filtered.substr(0, 5) == "wait ") && (filtered.find("wait for attack") == std::string::npos))
     {
         std::string remaining = filtered.substr(filtered.find(" ") + 1);
         uint32 delay = atof(remaining.c_str()) * IN_MILLISECONDS;
