+++ playerbot/strategy/actions/ChatActionContext.h
@@ -68,6 +68,7 @@
 #include "CheatAction.h"
 #include "GuildManagementActions.h"
 #include "RtscAction.h"
+#include "WaitForAttackAction.h"
 
 namespace ai
 {
@@ -159,6 +160,7 @@ namespace ai
             creators["rtsc"] = &ChatActionContext::rtsc;
             creators["ah"] = &ChatActionContext::ah;
             creators["ah bid"] = &ChatActionContext::ah_bid;
+            creators["wait for attack time"] = &ChatActionContext::wait_for_attack_time;
 
             creators["ginvite"] = &ChatActionContext::ginvite;
             creators["guild promote"] = &ChatActionContext::guild_promote;
@@ -253,6 +255,7 @@ namespace ai
         static Action* rtsc(PlayerbotAI* ai) { return new RTSCAction(ai); }
         static Action* ah(PlayerbotAI* ai) { return new AhAction(ai); }
         static Action* ah_bid(PlayerbotAI* ai) { return new AhBidAction(ai); }
+        static Action* wait_for_attack_time(PlayerbotAI* ai) { return new WaitForAttackSetTimeAction(ai); }
 
         static Action* ginvite(PlayerbotAI* ai) { return new GuildInviteAction(ai); }
         static Action* guild_promote(PlayerbotAI* ai) { return new GuildPromoteAction(ai); }
